Major Changes
1. Modified Severity Level to SEVERE in Base Asset
2. Removed Joint JS 0.9.6 Uncompressed Version
3. Removed Alphanum Dependency
4. Added Joint JS 0.9.7 Minimized Dependency
5. Refactored Wire Components Anchor List Item
6. Refactored FindConfigurationByPid in GwtComponentService
7. Added functionality to show error modal if user wants to add an item
after it deletes an instance from Wire Graph. For this, user needs to
save the graph first. This actually prevents the use case when user
deletes an old Wire Component instance and creates a new instance with
the same name. After clicking on the new instance, it shows the old
properties associated with the old instance. This happended due to the
fact the old instance was not still deleted from OSGi.
8. Added dummy click handler in Add Channel
9. WiresPanelUI refactored

Signed-off-by: Amit Kumar Mondal <admin@amitinside.com>